+++ packages/llm_dart_phind/lib/llm_dart_phind.dart
@@ -1,5 +1,0 @@
-library;
-
-export 'defaults.dart';
-export 'phind.dart';
-export 'phind_factory.dart';